DROOLS-2919: [DMN Designer] Data-types: Clean-up - Updates following peer review.
